+++ tests/CopperfinTestIsolation.cmake
@@ -178,7 +178,8 @@ function(copperfin_configure_native_test_isolation)
     endforeach()
 
     foreach(test_name IN ITEMS
-            test_platform_models)
+            test_platform_models
+            test_platform_safe_regex)
         copperfin_set_test_isolation(${test_name}
             PARALLEL_SAFE
             FILESYSTEM none
@@ -989,6 +990,7 @@ function(copperfin_configure_native_test_isolation)
             test_prg_engine_functions
             test_prg_engine_index_seek_optimization
             test_prg_engine_json_facade
+            test_prg_engine_regex_facade
             test_prg_engine_locale_code_page
             test_prg_engine_parser_classes
             test_prg_engine_path_functions